Tools Potential Risk refactoring and retranslating
