+++ app/components/ToolsPotentialRisk/Container.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+export default styled.div`
+  text-align: ${props=>props.theme.lang === 'ar' ? 'right' : 'left'};
+`;
