+++ app/components/ToolsPotentialRisk/HeaderName.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+export default styled.span`
+  margin-top: ${p=>p.theme.lang==='ar'?'-4px':'0'};
+`;
