+++ app/containers/ToolsPotentialRisk/staticText.js
@@ -1,0 +1,6 @@
+export default {
+  potentialRiskHeader: {
+    id: 'module.potential-risks',
+    defaultMessage: 'Potential Risks',
+  },
+}
